+++ Civi/ActionProvider/Action/CiviCase/GetCaseDataById.php
@@ -112,7 +112,7 @@ class GetCaseDataById extends AbstractAction {
             $type = 'Integer';
             break;
           case 'File':
-            $type = null;
+            $type = 'Integer';
             break;
           case 'Memo':
             $type = 'Text';